java backend: FUNCTION CACHING: Added support for Nailgun. Specifically: 1. Moved cache data structure into a separate class managed by globalContext, not static like it was before. 2. Moved cache logic from JavaSymbolicObject to Term. This is mostly architecture/performance optimization, because it is only used for Term.
